Fix dependency issue when trying to build and run
I was unable to build and run the spectrogram experiment due to the
missmatch in folder names and what was trying to be imported. I
found that it was simply an issue with casing. Not sure if the
environment this was built in has a case insensitive filesystem or
not, but I run Linux as my daily driver, and it definitely does :)